AI generated unit tests

+++ auth.go
@@ -88,7 +88,7 @@ func jwtMiddleware(next http.HandlerFunc) http.HandlerFunc {
 			return
 		}
 
-		r = r.WithContext(context.WithValue(r.Context(), ContextKeyUserID, uint64(claims.UserID)))
+		r = r.WithContext(context.WithValue(r.Context(), ContextKeyUserID, claims.UserID))
 		r = r.WithContext(context.WithValue(r.Context(), ContextKeyUsername, claims.Username))
 
 		r.Header.Set(AuthHeaderUserID, strconv.FormatUint(uint64(claims.UserID), 10))
